Updated links in the quotes section (#40)
The links in the quotes section were on the person's name, for example, "DR. BHASKARA MARTHI"
However if you clicked the link it directs you to Dave's Linkedin page. This was either a bug or it was intended that you would read the full quote on Dave's linkedin. In either case, I think it's very counter intuitive to click on someone's name and be directed to a different person's linkedin. Plus, they can read the full quote in the "references given" section of DR. BHASKARA MARTHI

Fixed the four links